+++ src/main/java/org/gridsuite/actions/server/ActionsApi.java
@@ -4,12 +4,15 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.actions.server.contingency;
+package org.gridsuite.actions.server;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public interface ContingencyList {
+public final class ActionsApi {
 
-    String getName();
+    private ActionsApi() {
+    }
+
+    public static final String API_VERSION = "v1";
 }
