+++ src/main/java/org/gridsuite/actions/server/ScriptContingencyListEntity.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.actions.server.contingency;
+package org.gridsuite.actions.server;
 
 import lombok.Getter;
 import org.springframework.data.cassandra.core.cql.PrimaryKeyType;
